requirements.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
alabaster==0.7.12
appdirs==1.4.4
arrow==1.1.1
astroid==2.6.5
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
bibtexparser==1.2.0
certifi==2021.5.30
cffi==1.14.6
cfgv==3.3.0
chardet==3.0.4
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
discord-components==1.1.4
discord-py-slash-command==2.3.2
discord.py==1.7.3
distlib==0.3.2
docutils==0.16
fake-useragent==0.1.11
filelock==3.0.12
free-proxy==1.0.2
future==0.18.2
identify==2.2.11
idna==3.2
imagesize==1.2.0
importlib-metadata==4.6.1
iso-639==0.4.5
isort==5.9.2
Jinja2==3.0.1
langid==1.1.6
lazy-object-proxy==1.6.0
libretranslatepy==2.1.1
lxml==4.6.3
mal-api==0.4.0
MarkupSafe==2.0.1
mccabe==0.6.1
multidict==4.7.6
nodeenv==1.6.0
numpy==1.21.1
packaging==21.0
pep517==0.11.0
Pillow==8.3.1
pip-tools==6.2.0
platformdirs==2.1.0
pluggy==0.13.1
pornhub-api==0.2.0
pre-commit==2.13.0
py==1.10.0
pycparser==2.20
pydantic==1.8.2
Pygments==2.9.0
pylint==2.9.6
pyOpenSSL==20.0.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytube==10.9.3
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
scholarly==1.2.2
selenium==3.141.0
six==1.16.0
snowballstemmer==2.1.0
soupsieve==2.2.1
Sphinx==4.1.2
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stem==1.8.0
toml==0.10.2
tomli==1.1.0
tox==3.24.0
translate==3.6.1
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.6
virtualenv==20.6.0
wikipedia==1.4.0
wrapt==1.12.1
yarl==1.6.3
youtube-search==2.1.0
zipp==3.5.0
